Style fixes and simplifications to TableList after review
